pyproject.toml
[tool.poetry]
name = "videsc"
version = "0.1.0"
description = "Video Description Service"
authors = ["Edward Fioletov <naereni12@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
tqdm = "^4.65.0"
pandas = "^1.5.3"
numpy = "^1.24.2"
opencv-python = "^4.7.0.72"
transformers = "^4.27.4"
wandb = "^0.14.0"
pillow = "^9.5.0"
torch = "^2.0.0"
nltk = "^3.8.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"
black = "^23.3.0"
mypy = "^1.1.1"
pre-commit = "^3.2.1"
jupyter = "^1.0.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79

[tool.black]
line-length = 79
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
